Add initial files for FinancialApp project
This commit adds the base files required to set up the FinancialApp project. It includes the index.html, index.js, config.xml and .gitignore files. Each of these files serve various roles, including setting up the primary URL loading point, the configurations for the application, and the files to be ignored during commits.